+++ Packages/idv.jlchntoz.vvmw/Runtime/VVMW/Core_Audio.cs
@@ -12,8 +12,12 @@ namespace JLChnToZ.VRC.VVMW {
         float defaultVolume = 1;
         [SerializeField, LocalizedLabel, FieldChangeCallback(nameof(Muted))]
         bool defaultMuted = false;
-        [SerializeField, LocalizedLabel] bool muteOnOutOfRange;
+        [SerializeField, LocalizedLabel, Min(0)] float volumeFadeDuration = 0.5f;
+        [SerializeField, LocalizedLabel] internal bool muteOnOutOfRange;
+        [SerializeField, LocalizedLabel, Range(0, 1)] internal float outOfRangeVolume = 0f;
         AudioSource assignedAudioSource, assignedAudioSourceR;
+        float currentFadeVolume = 0, targetFadeVolume = 0;
+        bool isVolumeFading = false;
 
         /// <summary>
         /// The volume of the audio.
@@ -41,34 +45,53 @@ namespace JLChnToZ.VRC.VVMW {
         }
 
         void UpdateVolume() {
-            var volume = defaultMuted || (muteOnOutOfRange && hasRegion && !isCurrentCoreActive) ? 0 : defaultVolume * defaultVolume; // Volume is not linear
+            if (!isVolumeFading) {
+                isVolumeFading = true;
+                _HandleVolumeFade();
+            }
+            SendEvent("_OnVolumeChange");
+            UpdateAudioLinkVolume();
+        }
+
+#if COMPILER_UDONSHARP
+        public
+#endif
+        void _HandleVolumeFade() {
+            targetFadeVolume = defaultMuted ? 0 : hasRegion && !isCurrentCoreActive ? outOfRangeVolume : 1F;
+            if (volumeFadeDuration <= 0 || Mathf.Approximately(currentFadeVolume, targetFadeVolume)) {
+                currentFadeVolume = targetFadeVolume;
+                isVolumeFading = false;
+            } else {
+                currentFadeVolume = Mathf.MoveTowards(currentFadeVolume, targetFadeVolume, UnityEngine.Time.deltaTime / volumeFadeDuration);
+                SendCustomEventDelayedFrames(nameof(_HandleVolumeFade), 0);
+            }
+            var volume = currentFadeVolume * defaultVolume;
+            volume *= volume; // Volume is not linear
             if (Utilities.IsValid(audioSources))
-                for (int i = 0; i < audioSources.Length; i++) {
+                for (int i = 0, ic = audioSources.Length; i < ic; i++) {
                     var audioSource = audioSources[i];
                     if (!Utilities.IsValid(audioSource)) continue;
                     audioSource.volume = volume;
                 }
             if (Utilities.IsValid(audioControllers))
-                for (int i = 0; i < audioControllers.Length; i++) {
+                for (int i = 0, ic = audioControllers.Length; i < ic; i++) {
                     var controller = audioControllers[i];
                     if (!Utilities.IsValid(controller)) continue;
                     controller.InternalSetVolume(volume);
                 }
-            SendEvent("_OnVolumeChange");
-            UpdateAudioLinkVolume();
         }
 
         void SetAudioPitch() {
             if (!Utilities.IsValid(activeHandler)) return;
             var speed = activeHandler.Speed;
             if (Utilities.IsValid(audioSources))
-                for (int i = 0; i < audioSources.Length; i++) {
+                for (int i = 0, ic = audioSources.Length; i < ic; i++) {
                     var audioSource = audioSources[i];
                     if (!Utilities.IsValid(audioSource)) continue;
                     audioSource.pitch = speed;
                 }
             if (Utilities.IsValid(audioControllers))
-                for (int i = 0; i < audioControllers.Length; i++) {
+                for (int i = 0, ic = audioControllers.Length; i < ic; i++) {
                     var controller = audioControllers[i];
                     if (!Utilities.IsValid(controller)) continue;
                     controller.InternalSetPitch(speed);